# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Concesionarios API",
    "version": "1.0.0",
    "description": "API para gestionar concesionarios y coches."
  },
  "basePath": "/",
  "definitions": {
    "Coche": {
      "type": "object",
      "properties": {
        "modelo": { "type": "string" },
        "cv": { "type": "integer" },
        "precio": { "type": "number" }
      },
      "required": ["modelo", "cv", "precio"]
    },
    "Concesionario": {
      "type": "object",
      "properties": {
        "nombre": { "type": "string" },
        "direccion": { "type": "string" },
        "coches": {
          "type": "array",
          "items": { "$ref": "#/definitions/Coche" }
        }
      },
      "required": ["nombre", "direccion"]
    }
  },
  "paths": {
    "/concesionarios": {
      "get": {
        "summary": "Obtener todos los concesionarios",
        "operationId": "getConcesionarios",
        "responses": {
          "200": {
            "description": "Lista de concesionarios",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Concesionario" }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo concesionario",
        "operationId": "createConcesionario",
        "parameters": [
          {
            "in": "body",
            "name": "concesionario",
            "schema": { "$ref": "#/definitions/Concesionario" }
          }
        ],
        "responses": {
          "200": { "description": "Concesionario creado" }
        }
      }
    },
    "/concesionarios/{id}": {
      "get": {
        "summary": "Obtener un concesionario por ID",
        "operationId": "getConcesionarioById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Concesionario obtenido",
            "schema": { "$ref": "#/definitions/Concesionario" }
          }
        }
      },
      "put": {
        "summary": "Actualizar un concesionario por ID",
        "operationId": "updateConcesionarioById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "concesionario",
            "schema": { "$ref": "#/definitions/Concesionario" }
          }
        ],
        "responses": {
          "200": { "description": "Concesionario actualizado" }
        }
      },
      "delete": {
        "summary": "Borrar un concesionario por ID",
        "operationId": "deleteConcesionarioById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Concesionario eliminado" }
        }
      }
    },
    "/concesionarios/{id}/coches": {
      "get": {
        "summary": "Obtener coches de un concesionario",
        "operationId": "getCochesByConcesionarioId",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Lista de coches",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Coche" }
            }
          }
        }
      },
      "post": {
        "summary": "Añadir un coche a un concesionario",
        "operationId": "addCocheToConcesionario",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "coche",
            "schema": { "$ref": "#/definitions/Coche" }
          }
        ],
        "responses": {
          "200": { "description": "Coche añadido" }
        }
      }
    },
    "/concesionarios/{id}/coches/{cocheId}": {
      "get": {
        "summary": "Obtener un coche por ID",
        "operationId": "getCocheById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          { "in": "path", "name": "cocheId", "required": true, "type": "string" }
        ],
        "responses": {
          "200": {
            "description": "Coche obtenido",
            "schema": { "$ref": "#/definitions/Coche" }
          }
        }
      },
      "put": {
        "summary": "Actualizar un coche por ID",
        "operationId": "updateCocheById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          { "in": "path", "name": "cocheId", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "coche",
            "schema": { "$ref": "#/definitions/Coche" }
          }
        ],
        "responses": {
          "200": { "description": "Coche actualizado" }
        }
      },
      "delete": {
        "summary": "Eliminar un coche por ID",
        "operationId": "deleteCocheById",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "type": "string" },
          { "in": "path", "name": "cocheId", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Coche eliminado" }
        }
      }
    }
  }
}